refactor: 一部をAstroコンポーネント化してアイランド部分を最小限に
