+++ src/components/Sidebar/type.ts
@@ -1,0 +1,7 @@
+export type SidebarItem = {
+  link: `/${string}`;
+  order: number;
+  title: string;
+  depth: number;
+  children: SidebarItem[];
+};
